// libs/user/src/lib/user/types/actions/friends-app-action.types.ts
export enum FrndsAppInitActionTypes {
  FRNDS_INIT = '[FRNDS APP INIT] FRNDS APP INIT',
  FRNDS_INIT_SUCCESS = '[FRNDS APP INIT] FRNDS INIT SUCCESS',
  FRNDS_INIT_ERROR = '[FRNDS APP INIT] FRNDS INIT ERROR',
}

export enum FrndsAppListActionTypes {
  FRNDS_LIST_INIT = '[FRNDS APP LIST] FRNDS LIST INIT',
  FRNDS_LIST_INIT_SUCCESS = '[FRNDS APP LIST] FRNDS LIST INIT SUCCESS',
  FRNDS_LIST_INIT_ERROR = '[FRNDS APP LIST] FRNDS LIST INIT ERROR',
}

export enum FrndsAppDetailActionTypes {
  FRNDS_DETAILS_INIT = '[FRNDS APP DETAILS] FRNDS DETAILS INIT',
  FRNDS_DETAILS_INIT_SUCCESS = '[FRNDS APP DETAILS] FRNDS DETAILS SUCCESS',
  FRNDS_DETAILS_INIT_ERROR = '[FRNDS APP DETAILS] FRNDS DETAILS ERROR',
  FRNDS_DETAILS_CREATE = '[FRNDS APP DETAILS] FRNDS DETAILS CREATE',
  FRNDS_DETAILS_CREATE_SUCCESS = '[FRNDS APP DETAILS] FRNDS DETAILS CREATE SUCCESS',
  FRNDS_DETAILS_CREATE_FAILURE = '[FRNDS APP DETAILS] FRNDS DETAILS CREATE FAILURE',
  FRNDS_DETAILS_UPDATE = '[FRNDS APP DETAILS] FRNDS DETAILS UPDATE',
  FRNDS_DETAILS_UPDATE_SUCCESS = '[FRNDS APP DETAILS] FRNDS DETAILS UPDATE SUCCESS',
  FRNDS_DETAILS_UPDATE_FAILURE = '[FRNDS APP DETAILS] FRNDS DETAILS UPDATE FAILURE',
  FRNDS_DETAILS_DELETE = '[FRNDS APP DETAILS] FRNDS DETAILS DELETE',
  FRNDS_DETAILS_DELETE_SUCCESS = '[FRNDS APP DETAILS] FRNDS DETAILS DELETE SUCCESS',
  FRNDS_DETAILS_DELETE_FAILURE = '[FRNDS APP DETAILS] FRNDS DETAILS DELETE FAILURE',
  FRNDS_DETAILS_CLEAR = '[FRNDS APP DETAILS] FRNDS DETAILS CLEAR',
  FRNDS_DETAILS_CLEAR_SUCCESS = '[FRNDS APP DETAILS] FRNDS DETAILS CLEAR SUCCESS',
  FRNDS_DETAILS_CLEAR_FAILURE = '[FRNDS APP DETAILS] FRNDS DETAILS CLEAR FAILURE',
}

export enum FrndsAppQueryActionTypes {
  FRNDS_QUERY = '[FRNDS APP QUERY] FRNDS QUERY',
  FRNDS_QUERY_SUCCESS = '[FRNDS APP QUERY] FRNDS QUERY SUCCESS',
  FRNDS_QUERY_ERROR = '[FRNDS APP QUERY] FRNDS QUERY ERROR',
}

export enum FrndsAppNewUserActionTypes {
  FRNDS_NEW_USER_CLICK = '[FRNDS APP NEW_USER] FRNDS NEW_USER_CLICK',
  FRNDS_NEW_USER_INIT = '[FRNDS APP NEW_USER] FRNDS NEW_USER INIT',
  FRNDS_NEW_USER_SUCCESS = '[FRNDS APP NEW_USER] FRNDS NEW_USER SUCCESS',
  FRNDS_NEW_USER_ERROR = '[FRNDS APP NEW_USER] FRNDS NEW_USER ERROR',
}